+++ src/main/java/com/starkbank/Event.java
@@ -36,6 +36,7 @@ public class Event extends Resource {
 
     static {
         GsonEvent.registerTypeAdapter(Event.class, new Event.Deserializer());
+        GsonEvent.registerTypeAdapter(PaymentRequest.class, new PaymentRequest.Deserializer());
     }
 
     public String created;
@@ -116,6 +117,8 @@ public class Event extends Resource {
                     case "verified-account":
                         return context.deserialize(jsonObject,
                                 VerifiedAccountEvent.class);
+                    case "payment-request":
+                        return GsonEvent.getInstance().fromJson(jsonObject, PaymentRequestEvent.class);
                     default:
                         return context.deserialize(jsonObject,
                                 UnknownEvent.class);
@@ -290,6 +293,19 @@ public class Event extends Resource {
         }
 
         public VerifiedAccountEvent() {
+            super();
+        }
+    }
+
+    public final static class PaymentRequestEvent extends Event {
+        public PaymentRequest.Log log;
+
+        public PaymentRequestEvent(PaymentRequest.Log log, String created, Boolean isDelivered, String subscription, String id, String workspaceId) {
+            super(created, isDelivered, subscription, id, workspaceId);
+            this.log = log;
+        }
+
+        public PaymentRequestEvent() {
             super();
         }
     }