+++ src/test/java/TestPaymentRequest.java
@@ -4,6 +4,9 @@ import com.starkbank.Transaction;
 import com.starkbank.error.InputErrors;
 import com.starkbank.utils.Generator;
 import com.starkbank.utils.Resource;
+import com.starkbank.Event;
+import com.starkcore.error.InvalidSignatureError;
+import com.starkbank.BoletoPayment;
 import org.junit.Assert;
 import org.junit.AssumptionViolatedException;
 import org.junit.Test;
@@ -87,6 +90,27 @@ public class TestPaymentRequest {
         if (ids.size() != 4) {
             throw new Exception("ids.size() != 4");
         }
+    }
+
+    @Test
+    public void testPaymentRequestEventParse() throws Exception{
+        Settings.user = utils.User.defaultProject();
+
+        String content = "{\"event\": {\"created\": \"2025-12-02T19:26:10.088671+00:00\", \"id\": \"4954566443401216\", \"log\": {\"created\": \"2025-12-02T19:26:07.449707+00:00\", \"errors\": [], \"id\": \"5602863480832000\", \"request\": {\"actions\": [{\"action\": \"requested\", \"id\": \"6200062581407744\", \"type\": \"project\"}, {\"action\": \"required\", \"id\": \"6263563236671488\", \"type\": \"member\"}], \"amount\": 4000, \"centerId\": \"6315803154579456\", \"created\": \"2025-12-02T19:26:07.396749+00:00\", \"description\": \"Payment for killing white walkers\", \"due\": \"2025-12-07T03:00:00+00:00\", \"id\": \"6207755400511488\", \"payment\": {\"amount\": 4000, \"description\": \"Payment for killing white walkers\", \"line\": \"34191.09107 83495.507309 71444.640008 4 12970000004000\", \"name\": \"Iron Bank S.A.\", \"tags\": [\"little girl\", \"no one\"], \"taxId\": \"38.435.677/0001-25\"}, \"status\": \"pending\", \"tags\": [], \"type\": \"boleto-payment\", \"updated\": \"2025-12-02T19:26:07.449725+00:00\"}, \"type\": \"created\", \"updated\": \"2025-12-02T19:26:07.449729+00:00\", \"user\": {\"id\": \"6200062581407744\", \"type\": \"project\"}}, \"subscription\": \"payment-request\", \"workspaceId\": \"6341320293482496\"}}";
+        String valid_signature = "MEYCIQCeKDA2cQgiIv4oK9u4Hi3zbryBwZfc5gF81G2rsMlU4wIhANZXNnp1umRiERsIVgOQu0Z8thnJQKRBd22p4ldzuEV5";
+
+        Event.PaymentRequestEvent event = null;
+
+        try {
+            event = (Event.PaymentRequestEvent) Event.parse(content, valid_signature);
+        } catch (InvalidSignatureError e) {
+            Assert.fail(e.getMessage());
+        }
+
+        Assert.assertEquals(event.getClass(), Event.PaymentRequestEvent.class);
+        Assert.assertEquals(event.log.getClass(), PaymentRequest.Log.class);
+        Assert.assertEquals(event.log.request.getClass(), PaymentRequest.class);
+        Assert.assertEquals(event.log.request.payment.getClass(), BoletoPayment.class);
     }
 
     static PaymentRequest example() throws Exception{